Moved NotificationManager to Modules
