+++ source/Modules/NotificationManager.js
