+++ source/Modules/index.js
@@ -1,1 +1,2 @@
 export * from './storageManager';
+export { default as NotificationManager } from './NotificationManager';
